sensors/imu_driver.py
import time

class IMUDriver:
    def __init__(self, bus_number=1, address=0x68):
        """
        Initialize the IMU (e.g., MPU6050 or BNO055).
        For this skeleton, we assume an I2C connection on a Raspberry Pi.
        """
        self.address = address
        self.bus_number = bus_number
        self.connected = False
        self._setup()

    def _setup(self):
        try:
            import smbus2
            self.bus = smbus2.SMBus(self.bus_number)
            self.connected = True
            print(f"IMU initialized on I2C bus {self.bus_number} at {hex(self.address)}")
            # Write configuration registers here...
        except ImportError:
            print("smbus2 not found. IMU running in mock mode.")
            self.connected = False
        except Exception as e:
            print(f"Failed to initialize IMU: {e}")
            self.connected = False

    def get_data(self):
        """
        Read data from the IMU.
        Returns:
            dict: Containing 'accel_x', 'accel_y', 'accel_z', 'gyro_x', 'gyro_y', 'gyro_z', 'yaw_rate'
        """
        if not self.connected:
            # Mock data
            return {
                'accel_x': 0.0,
                'accel_y': 0.0,
                'accel_z': 9.81,
                'gyro_x': 0.0,
                'gyro_y': 0.0,
                'gyro_z': 0.0,
                'yaw_rate': 0.0, # radians/sec
                'timestamp': time.time()
            }
            
        try:
            # Read actual I2C registers here
            # For example, read 14 bytes starting from register 0x3B for MPU6050
            # block = self.bus.read_i2c_block_data(self.address, 0x3B, 14)
            # Parse the bytes into floats...
            
            # Placeholder return
            return {
                'accel_x': 0.0,
                'accel_y': 0.0,
                'accel_z': 9.81,
                'gyro_x': 0.0,
                'gyro_y': 0.0,
                'gyro_z': 0.0,
                'yaw_rate': 0.0,
                'timestamp': time.time()
            }
        except Exception as e:
            print(f"Error reading IMU: {e}")
            return None
