requirements.txt
numpy==1.26.4
scipy==1.12.0
filterpy==1.4.5
rplidar-robotic==1.0.0
pyserial==3.5
# If using GPIO on RPi, we might use RPi.GPIO or gpiozero
# RPi.GPIO==0.7.1
# smbus2==0.4.3  # For I2C communication (IMU)
